app: add Modal component, replace Backdrop
