+++ app/src/App.svelte
@@ -1,16 +1,14 @@
 <script lang="ts">
 	import TopBar from '$lib/Components/TopBar.svelte';
-	import Disclaimer from '$lib/Components/Disclaimer.svelte';
-	import Settings from '$lib/Components/Settings.svelte';
 	import Launch from '$lib/Components/Launch.svelte';
 	import Auth from '$lib/Components/Auth.svelte';
 	import PluginMenu from '$lib/Components/PluginMenu.svelte';
 	import { urlSearchParams } from '$lib/Util/functions';
-	import { showDisclaimer } from '$lib/Util/store';
 	import LogView from '$lib/Components/LogView.svelte';
 	import { logger } from '$lib/Util/Logger';
+	import { showDisclaimer } from '$lib/Util/store';
+	import DisclaimerModal from '$lib/Components/DisclaimerModal.svelte';
 
-	let showSettings: boolean = false;
 	let showPluginMenu: boolean = false;
 	let authorizing: boolean = false;
 
@@ -55,13 +53,11 @@
 		{#if showPluginMenu}
 			<PluginMenu bind:showPluginMenu></PluginMenu>
 		{/if}
-		{#if showSettings}
-			<Settings bind:showSettings></Settings>
+		{#if showDisclaimer}
+			<!-- TODO: remove showDisclaimer -->
+			<DisclaimerModal></DisclaimerModal>
 		{/if}
-		{#if $showDisclaimer}
-			<Disclaimer></Disclaimer>
-		{/if}
-		<TopBar bind:showSettings></TopBar>
+		<TopBar></TopBar>
 		<div class="mt-16 grid h-full grid-flow-col grid-cols-3">
 			<div></div>
 			<Launch bind:showPluginMenu></Launch>
