+++ app/src/lib/Components/PluginMenu.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
 	import { get } from 'svelte/store';
 	import { onDestroy } from 'svelte';
-	import Backdrop from '$lib/Components/Backdrop.svelte';
 	import { getNewClientListPromise, savePluginConfig } from '$lib/Util/functions';
 	import { type PluginConfig } from '$lib/Util/interfaces';
 	import { clientListPromise, hasBoltPlugins, pluginList, platform } from '$lib/Util/store';
 	import { logger } from '$lib/Util/Logger';
+	import Modal from '$lib/Components/CommonUI/Modal.svelte';
 
 	// props
 	export let showPluginMenu: boolean;
@@ -136,148 +136,137 @@
 	});
 </script>
 
-<div>
-	<Backdrop on:click={tryExit}></Backdrop>
+<Modal on:close={tryExit} class="h-[90%] w-[90%]">
 	<div
-		class="absolute left-[5%] top-[5%] z-20 h-[90%] w-[90%] rounded-lg bg-slate-100 text-center shadow-lg dark:bg-slate-900"
+		class="left-0 float-left h-full w-[min(180px,_50%)] overflow-hidden border-r-2 border-slate-300 pt-2 dark:border-slate-800"
 	>
 		<button
-			class="absolute right-3 top-3 rounded-full bg-rose-500 p-[2px] shadow-lg duration-200 hover:rotate-90 hover:opacity-75"
-			on:click={tryExit}
+			class="mx-auto mb-2 w-[95%] rounded-lg border-2 {isClientSelected
+				? 'border-blue-500 text-black dark:text-white'
+				: 'border-black bg-blue-500 text-black'} p-2 font-bold hover:opacity-75"
+			on:click={() => (isClientSelected = false)}
 		>
-			<img src="svgs/xmark-solid.svg" class="h-5 w-5" alt="Close" />
+			Manage Plugins
 		</button>
-		<div
-			class="left-0 float-left h-full w-[min(180px,_50%)] overflow-hidden border-r-2 border-slate-300 pt-2 dark:border-slate-800"
-		>
-			<button
-				class="mx-auto mb-2 w-[95%] rounded-lg border-2 {isClientSelected
-					? 'border-blue-500 text-black dark:text-white'
-					: 'border-black bg-blue-500 text-black'} p-2 font-bold hover:opacity-75"
-				on:click={() => (isClientSelected = false)}
-			>
-				Manage Plugins
-			</button>
-			<hr class="p-1 dark:border-slate-700" />
-			{#await $clientListPromise}
-				<p>loading...</p>
-			{:then clients}
-				{#if clients.length == 0}
-					<p>(start an RS3 game client with plugins enabled and it will be listed here.)</p>
-				{:else}
-					{#each clients as client}
-						<button
-							on:click={() => {
-								selectedClientId = client.uid;
-								isClientSelected = true;
-							}}
-							class="m-1 h-[28px] w-[95%] rounded-lg border-2 {isClientSelected &&
-							selectedClientId === client.uid
-								? 'border-black bg-blue-500 text-black'
-								: 'border-blue-500 text-black dark:text-white'} hover:opacity-75"
-						>
-							{client.identity || unnamedClientName}
-						</button>
-						<br />
-					{/each}
-				{/if}
-			{:catch}
-				<p>error</p>
-			{/await}
-		</div>
-		<div class="h-full pt-10">
-			{#if hasBoltPlugins}
-				<select
-					bind:value={selectedPlugin}
-					class="mx-auto mb-4 w-[min(280px,_45%)] cursor-pointer rounded-lg border-2 border-slate-300 bg-inherit p-2 text-inherit duration-200 hover:opacity-75 dark:border-slate-800"
-				>
-					{#each Object.entries($pluginList) as [id, plugin]}
-						<option class="dark:bg-slate-900" value={id}>{plugin.name ?? unnamedPluginName}</option>
-					{/each}
-				</select>
-				{#if !isClientSelected}
+		<hr class="p-1 dark:border-slate-700" />
+		{#await $clientListPromise}
+			<p>loading...</p>
+		{:then clients}
+			{#if clients.length == 0}
+				<p>(start an RS3 game client with plugins enabled and it will be listed here.)</p>
+			{:else}
+				{#each clients as client}
 					<button
-						class="aspect-square w-8 rounded-lg border-2 border-blue-500 text-[20px] font-bold duration-200 enabled:hover:opacity-75 disabled:border-gray-500"
-						on:click={jsonFilePicker}
-						disabled={disableButtons}
+						on:click={() => {
+							selectedClientId = client.uid;
+							isClientSelected = true;
+						}}
+						class="m-1 h-[28px] w-[95%] rounded-lg border-2 {isClientSelected &&
+						selectedClientId === client.uid
+							? 'border-black bg-blue-500 text-black'
+							: 'border-blue-500 text-black dark:text-white'} hover:opacity-75"
 					>
-						+
+						{client.identity || unnamedClientName}
 					</button>
 					<br />
-					{#if Object.entries($pluginList).length !== 0}
-						{#if Object.keys(get(pluginList)).includes(selectedPlugin) && managementPluginPromise !== null}
-							{#await managementPluginPromise}
-								<p>loading...</p>
-							{:then plugin}
-								<p class="pb-4 text-xl font-bold">
-									{plugin.name ?? unnamedPluginName}
-								</p>
-								<p class={plugin.description ? null : 'italic'}>
-									{plugin.description ?? 'no description'}
-								</p>
-								<br />
-							{:catch}
-								<p>error</p>
-								<br />
-							{/await}
-							<button
-								class="mx-auto mb-1 w-[min(144px,_25%)] rounded-lg p-2 font-bold text-black duration-200 enabled:bg-rose-500 enabled:hover:opacity-75 disabled:bg-gray-500"
-								on:click={() => {
-									managementPluginPromise = null;
-									pluginConfigDirty = true;
-									let list = get(pluginList);
-									delete list[selectedPlugin];
-									pluginList.set(list);
-								}}
-							>
-								Remove
-							</button>
-							<button
-								class="mx-auto mb-1 w-[min(144px,_25%)] rounded-lg p-2 font-bold text-black duration-200 enabled:bg-blue-500 enabled:hover:opacity-75 disabled:bg-gray-500"
-								on:click={() =>
-									(managementPluginPromise = getPluginConfigPromiseFromID(selectedPlugin))}
-							>
-								Reload
-							</button>
-						{/if}
-					{:else}
-						<p>
-							You have no plugins installed. Click the + button and select a plugin's bolt.json file
-							to add it.
-						</p>
+				{/each}
+			{/if}
+		{:catch}
+			<p>error</p>
+		{/await}
+	</div>
+	<div class="h-full pt-10">
+		{#if hasBoltPlugins}
+			<select
+				bind:value={selectedPlugin}
+				class="mx-auto mb-4 w-[min(280px,_45%)] cursor-pointer rounded-lg border-2 border-slate-300 bg-inherit p-2 text-inherit duration-200 hover:opacity-75 dark:border-slate-800"
+			>
+				{#each Object.entries($pluginList) as [id, plugin]}
+					<option class="dark:bg-slate-900" value={id}>{plugin.name ?? unnamedPluginName}</option>
+				{/each}
+			</select>
+			{#if !isClientSelected}
+				<button
+					class="aspect-square w-8 rounded-lg border-2 border-blue-500 text-[20px] font-bold duration-200 enabled:hover:opacity-75 disabled:border-gray-500"
+					on:click={jsonFilePicker}
+					disabled={disableButtons}
+				>
+					+
+				</button>
+				<br />
+				{#if Object.entries($pluginList).length !== 0}
+					{#if Object.keys(get(pluginList)).includes(selectedPlugin) && managementPluginPromise !== null}
+						{#await managementPluginPromise}
+							<p>loading...</p>
+						{:then plugin}
+							<p class="pb-4 text-xl font-bold">
+								{plugin.name ?? unnamedPluginName}
+							</p>
+							<p class={plugin.description ? null : 'italic'}>
+								{plugin.description ?? 'no description'}
+							</p>
+							<br />
+						{:catch}
+							<p>error</p>
+							<br />
+						{/await}
+						<button
+							class="mx-auto mb-1 w-[min(144px,_25%)] rounded-lg p-2 font-bold text-black duration-200 enabled:bg-rose-500 enabled:hover:opacity-75 disabled:bg-gray-500"
+							on:click={() => {
+								managementPluginPromise = null;
+								pluginConfigDirty = true;
+								let list = get(pluginList);
+								delete list[selectedPlugin];
+								pluginList.set(list);
+							}}
+						>
+							Remove
+						</button>
+						<button
+							class="mx-auto mb-1 w-[min(144px,_25%)] rounded-lg p-2 font-bold text-black duration-200 enabled:bg-blue-500 enabled:hover:opacity-75 disabled:bg-gray-500"
+							on:click={() =>
+								(managementPluginPromise = getPluginConfigPromiseFromID(selectedPlugin))}
+						>
+							Reload
+						</button>
 					{/if}
 				{:else}
-					<br />
-					{#await managementPluginPromise}
-						<p>loading...</p>
-					{:then plugin}
-						{#if plugin && plugin.main && Object.keys($pluginList).includes(selectedPlugin)}
-							{#if $pluginList[selectedPlugin].path}
-								<button
-									class="mx-auto mb-1 w-auto rounded-lg bg-emerald-500 p-2 font-bold text-black duration-200 hover:opacity-75"
-									on:click={() =>
-										startPlugin(
-											selectedClientId,
-											selectedPlugin,
-											$pluginList[selectedPlugin].path ?? '',
-											plugin.main ?? ''
-										)}
-								>
-									Start {plugin.name}
-								</button>
-							{:else}
-								<p>can't start plugin: no path is configured</p>
-							{/if}
-						{:else}
-							<p>can't start plugin: does not appear to be configured</p>
-						{/if}
-					{:catch}
-						<p>error</p>
-					{/await}
+					<p>
+						You have no plugins installed. Click the + button and select a plugin's bolt.json file
+						to add it.
+					</p>
 				{/if}
 			{:else}
-				<p>error</p>
+				<br />
+				{#await managementPluginPromise}
+					<p>loading...</p>
+				{:then plugin}
+					{#if plugin && plugin.main && Object.keys($pluginList).includes(selectedPlugin)}
+						{#if $pluginList[selectedPlugin].path}
+							<button
+								class="mx-auto mb-1 w-auto rounded-lg bg-emerald-500 p-2 font-bold text-black duration-200 hover:opacity-75"
+								on:click={() =>
+									startPlugin(
+										selectedClientId,
+										selectedPlugin,
+										$pluginList[selectedPlugin].path ?? '',
+										plugin.main ?? ''
+									)}
+							>
+								Start {plugin.name}
+							</button>
+						{:else}
+							<p>can't start plugin: no path is configured</p>
+						{/if}
+					{:else}
+						<p>can't start plugin: does not appear to be configured</p>
+					{/if}
+				{:catch}
+					<p>error</p>
+				{/await}
 			{/if}
-		</div>
-	</div>
-</div>
+		{:else}
+			<p>error</p>
+		{/if}
+	</div></Modal
+>
